이름 및 url 수정
